Refactor website metrics components
