+++ src/app/(main)/websites/[websiteId]/WebsiteMetricsBar.tsx
@@ -1,21 +1,14 @@
-import { Dropdown, Item } from 'react-basics';
-import classNames from 'classnames';
-import { useDateRange, useMessages, useSticky } from '@/components/hooks';
-import WebsiteDateFilter from '@/components/input/WebsiteDateFilter';
+import { useDateRange, useMessages } from '@/components/hooks';
+import useWebsiteStats from '@/components/hooks/queries/useWebsiteStats';
 import MetricCard from '@/components/metrics/MetricCard';
 import MetricsBar from '@/components/metrics/MetricsBar';
-import { formatShortTime, formatLongNumber } from '@/lib/format';
-import useWebsiteStats from '@/components/hooks/queries/useWebsiteStats';
-import useStore, { setWebsiteDateCompare } from '@/store/websites';
-import WebsiteFilterButton from './WebsiteFilterButton';
-import styles from './WebsiteMetricsBar.module.css';
+import { formatLongNumber, formatShortTime } from '@/lib/format';
+import useStore from '@/store/websites';
 
 export function WebsiteMetricsBar({
   websiteId,
-  sticky,
   showChange = false,
   compareMode = false,
-  showFilter = false,
 }: {
   websiteId: string;
   sticky?: boolean;
@@ -26,7 +19,6 @@ export function WebsiteMetricsBar({
   const { dateRange } = useDateRange(websiteId);
   const { formatMessage, labels } = useMessages();
   const dateCompare = useStore(state => state[websiteId]?.dateCompare);
-  const { ref, isSticky } = useSticky({ enabled: sticky });
   const { data, isLoading, isFetched, error } = useWebsiteStats(
     websiteId,
     compareMode && dateCompare,
@@ -76,60 +68,24 @@ export function WebsiteMetricsBar({
       ]
     : [];
 
-  const items = [
-    { label: formatMessage(labels.previousPeriod), value: 'prev' },
-    { label: formatMessage(labels.previousYear), value: 'yoy' },
-  ];
-
   return (
-    <div
-      ref={ref}
-      className={classNames(styles.container, {
-        [styles.sticky]: sticky,
-        [styles.isSticky]: sticky && isSticky,
+    <MetricsBar isLoading={isLoading} isFetched={isFetched} error={error}>
+      {metrics.map(({ label, value, prev, change, formatValue, reverseColors }) => {
+        return (
+          <MetricCard
+            key={label}
+            value={value}
+            previousValue={prev}
+            label={label}
+            change={change}
+            formatValue={formatValue}
+            reverseColors={reverseColors}
+            showChange={!isAllTime && (compareMode || showChange)}
+            showPrevious={!isAllTime && compareMode}
+          />
+        );
       })}
-    >
-      <div>
-        <MetricsBar isLoading={isLoading} isFetched={isFetched} error={error}>
-          {metrics.map(({ label, value, prev, change, formatValue, reverseColors }) => {
-            return (
-              <MetricCard
-                key={label}
-                value={value}
-                previousValue={prev}
-                label={label}
-                change={change}
-                formatValue={formatValue}
-                reverseColors={reverseColors}
-                showChange={!isAllTime && (compareMode || showChange)}
-                showPrevious={!isAllTime && compareMode}
-              />
-            );
-          })}
-        </MetricsBar>
-      </div>
-      <div className={styles.actions}>
-        {showFilter && <WebsiteFilterButton websiteId={websiteId} />}
-        <WebsiteDateFilter websiteId={websiteId} showAllTime={!compareMode} />
-        {compareMode && (
-          <div className={styles.vs}>
-            <b>VS</b>
-            <Dropdown
-              className={styles.dropdown}
-              items={items}
-              value={dateCompare || 'prev'}
-              renderValue={value => items.find(i => i.value === value)?.label}
-              alignment="end"
-              onChange={(value: any) => setWebsiteDateCompare(websiteId, value)}
-            >
-              {items.map(({ label, value }) => (
-                <Item key={value}>{label}</Item>
-              ))}
-            </Dropdown>
-          </div>
-        )}
-      </div>
-    </div>
+    </MetricsBar>
   );
 }
