+++ src/app/(main)/websites/[websiteId]/events/EventsPage.tsx
@@ -1,14 +1,15 @@
 'use client';
+import { useMessages } from '@/components/hooks';
+import { GridRow } from '@/components/layout/Grid';
+import EventsChart from '@/components/metrics/EventsChart';
+import MetricsTable from '@/components/metrics/MetricsTable';
+import { useState } from 'react';
+import { Item, Tabs } from 'react-basics';
 import WebsiteHeader from '../WebsiteHeader';
+import { WebsiteMetrics } from '../WebsiteMetrics';
+import EventProperties from './EventProperties';
 import EventsDataTable from './EventsDataTable';
 import EventsMetricsBar from './EventsMetricsBar';
-import EventsChart from '@/components/metrics/EventsChart';
-import { GridRow } from '@/components/layout/Grid';
-import MetricsTable from '@/components/metrics/MetricsTable';
-import { useMessages } from '@/components/hooks';
-import { Item, Tabs } from 'react-basics';
-import { useState } from 'react';
-import EventProperties from './EventProperties';
 
 export default function EventsPage({ websiteId }) {
   const [tab, setTab] = useState('activity');
@@ -17,7 +18,9 @@ export default function EventsPage({ websiteId }) {
   return (
     <>
       <WebsiteHeader websiteId={websiteId} />
-      <EventsMetricsBar websiteId={websiteId} />
+      <WebsiteMetrics websiteId={websiteId}>
+        <EventsMetricsBar websiteId={websiteId} />
+      </WebsiteMetrics>
       <GridRow columns="two-one">
         <EventsChart websiteId={websiteId} />
         <MetricsTable
